requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain==0.1.0
openai==1.12.0
faiss-cpu==1.7.4
PyYAML==6.0.1
python-dotenv==1.0.0